+++ docs/CHANGELOG.md
@@ -105,6 +105,7 @@ Version 1.3.0 (Bakeapple): Not yet released
 - Add `require_d3d11` to dedicated server config
 - Add `$Stock Alpha Test` to `af_level_quirks.tbl` and use to restore stock alpha test only on known affected levels
 - Serve player type, team, and score in `players_request` packet response (for online RFSB)
+- Support 8-bit greyscale TGA textures (types 3 and 11) in editor and game
 
 [@is-this-c](https://github.com/is-this-c)
 - Use 64-bit integers for time deltas